requirements.txt
sounddevice
numpy
requests
python-dotenv
